Fix/profil img (#24)
* fin

* finfin
